split kubeclient
